requirements.txt
altair
asttokens
attrs
backcall
beautifulsoup4
bleach
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
decorator
defusedxml
docopt
executing
fastjsonschema
fonttools
gitdb
GitPython
idna
importlib_metadata
importlib_resources
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
narwhals
nbclient
nbconvert
nbformat
numpy
packaging
pandas
pandocfilters
parso
pickleshare
pillow
pipreqs
pkgutil_resolve_name
platformdirs
plotly
prompt_toolkit
protobuf
pure_eval
pyarrow
pydeck
Pygments
pyparsing
python-dateutil
pytz
pywin32
pyzmq
referencing
requests
rich
rpds-py
scikit-learn
scipy
seaborn
six
smmap
soupsieve
stack-data
streamlit
tenacity
threadpoolctl
tinycss2
toml
tornado
traitlets
typing_extensions
tzdata
urllib3
watchdog
wcwidth
webencodings
yarg
zipp
